+++ aprsd/cli_helper.py
@@ -1,4 +1,72 @@
+from functools import update_wrapper
+import typing as t
+
 import click
+
+from aprsd import config as aprsd_config
+from aprsd import log
+
+
+F = t.TypeVar("F", bound=t.Callable[..., t.Any])
+
+common_options = [
+    click.option(
+        "--loglevel",
+        default="INFO",
+        show_default=True,
+        type=click.Choice(
+            ["CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG"],
+            case_sensitive=False,
+        ),
+        show_choices=True,
+        help="The log level to use for aprsd.log",
+    ),
+    click.option(
+        "-c",
+        "--config",
+        "config_file",
+        show_default=True,
+        default=aprsd_config.DEFAULT_CONFIG_FILE,
+        help="The aprsd config file to use for options.",
+    ),
+    click.option(
+        "--quiet",
+        is_flag=True,
+        default=False,
+        help="Don't log to stdout",
+    ),
+]
+
+
+def add_options(options):
+    def _add_options(func):
+        for option in reversed(options):
+            func = option(func)
+        return func
+    return _add_options
+
+
+def process_standard_options(f: F) -> F:
+    def new_func(*args, **kwargs):
+        print(f"ARGS {args}")
+        print(f"KWARGS {kwargs}")
+        ctx = args[0]
+        ctx.ensure_object(dict)
+        ctx.obj["loglevel"] = kwargs["loglevel"]
+        ctx.obj["config_file"] = kwargs["config_file"]
+        ctx.obj["quiet"] = kwargs["quiet"]
+        ctx.obj["config"] = aprsd_config.parse_config(kwargs["config_file"])
+        log.setup_logging(
+            ctx.obj["config"], ctx.obj["loglevel"],
+            ctx.obj["quiet"],
+        )
+
+        del kwargs["loglevel"]
+        del kwargs["config_file"]
+        del kwargs["quiet"]
+        return f(*args, **kwargs)
+
+    return update_wrapper(t.cast(F, new_func), f)
 
 
 class AliasedGroup(click.Group):
@@ -33,42 +101,3 @@ class AliasedGroup(click.Group):
                 self.add_command(cmd, name=alias)
             return cmd
         return decorator
-
-
-class GroupWithCommandOptions(click.Group):
-    """ Allow application of options to group with multi command """
-
-    def add_command(self, cmd, name=None):
-        click.Group.add_command(self, cmd, name=name)
-
-        # add the group parameters to the command
-        for param in self.params:
-            cmd.params.append(param)
-
-        # hook the commands invoke with our own
-        cmd.invoke = self.build_command_invoke(cmd.invoke)
-        self.invoke_without_command = True
-
-    def build_command_invoke(self, original_invoke):
-
-        def command_invoke(ctx):
-            """ insert invocation of group function """
-
-            # separate the group parameters
-            ctx.obj = dict(_params={})
-            for param in self.params:
-                name = param.name
-                if name in ctx.params:
-                    ctx.obj["_params"][name] = ctx.params[name]
-                    del ctx.params[name]
-
-            # call the group function with its parameters
-            params = ctx.params
-            ctx.params = ctx.obj["_params"]
-            self.invoke(ctx)
-            ctx.params = params
-
-            # now call the original invoke(the command)
-            original_invoke(ctx)
-
-        return command_invoke
